Moved config to root directory
